implement authentication using the options pattern
it was not obvious for any user of this library to see how to send
authenticated requests to the ntfy server. I initially implemented a
custom http client with custom roundtripper when I discovered that the
Header object was public in the publisher. A clearer and more idiomatic
pattern is to use options to specify a custom http client and/or auth
credentials. This is what is implemented here.